Feat/#13 feat 이번달 전력량 및 목표량 조회 구현

+++ src/main/java/live/smoothing/sensordata/dto/goal/GoalRequest.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GoalRequest {
-    private Long goalAmount;
+    private Double goalAmount;
     private Integer unitPrice;
 }
